+++ crates/ic-query/src/lib.rs
@@ -16,7 +16,6 @@
 
 #[cfg(feature = "subnet-catalog-host")]
 mod agent;
-#[cfg(feature = "host")]
 pub mod cache;
 #[cfg(feature = "subnet-catalog-host")]
 mod cache_file;