+++ crates/ic-query/src/sns/report/text/proposals/cache_list.rs
@@ -34,7 +34,7 @@ pub fn sns_proposals_cache_list_report_text(report: &SnsCacheListReport) -> Stri
                 .iter()
                 .map(|cache| {
                     [
-                        cache.cache_status.clone(),
+                        cache.cache_status.to_string(),
                         cache.id.to_string(),
                         cache.name.clone(),
                         cache.root_canister_id.clone(),